% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implvol.R
\docType{package}
\name{implvol-package}
\alias{implvol}
\alias{implvol-package}
\title{Option implied volatility}
\description{
The implvol package computes option implied volatilities using Peter
Jaeckel's implementation of "Let's Be Rational".
}
\references{
\url{http://www.jaeckel.org/}
}
\seealso{
\code{\link[implvol]{implvol}}
}
\author{
\strong{Maintainer}: Chris Hein \email{cmhein00@gmail.com}

Authors:
\itemize{
  \item Peter Jaeckel
}

}
