% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implvol.R
\docType{package}
\name{implvol}
\alias{implvol}
\alias{implvol-package}
\title{implvol: option implied volatility}
\description{
implvol computes option implied volatilities using Peter Jaeckel's
implementation of "Let's Be Rational".
}
\author{
Chris Hein <cmhein00@gmail.com>
}
